requirements.txt
flask
gunicorn
google-cloud-datastore
google-cloud-translate
google-cloud-texttospeech==0.2.0
google-cloud-storage
six
